make is_gleam_module private again
