+++ compiler-cli/src/main.rs
@@ -64,7 +64,6 @@ mod fs;
 mod hex;
 mod http;
 mod lsp;
-mod module;
 mod new;
 mod panic;
 mod publish;
